# Optimise the XpathParser (time and recusion). 